cronctl: send log files via mail
it's always annoying to get an e-mail with an error and first have to
log into the server, to see whats going on.

This change refactors logging so that the log for each run can be found
in $(date)/failed or $(date)/succeeded.

The log for each run will also be send by e-mail.

Signed-off-by: Julian Heuking <J.Heuking@beckhoff.com>